+++ api/bro_import/object_import.py
@@ -3,6 +3,7 @@ import xml.etree.ElementTree as ET
 from abc import ABC, abstractmethod
 from typing import IO, Any
 
+import polars as pl
 import requests
 import xmltodict
 from django.conf import settings
@@ -82,15 +83,46 @@ class GMNObjectImporter(ObjectImporter):
         if "GMN_PPO" not in dispatch_document_data:
             return
 
-        gmn_data, measuringpoint_data = self._split_json_data(dispatch_document_data)
+        gmn_data, measuringpoint_data, intermediate_events = self._split_json_data(
+            dispatch_document_data
+        )
 
         self._save_gmn_data(gmn_data)
+        self._create_events_df(intermediate_events)
         self._save_measuringpoint_data(measuringpoint_data)
+
+    def _create_events_df(self, events_data: list[dict, Any]) -> None:
+        event_types = []
+        event_dates = []
+        measuring_point_codes = []
+        for event in events_data:
+            event_type = event.get("eventName", {}).get("#text", None)
+            event_date = event.get("eventDate", {}).get("brocom:date", None)
+            measuring_point_code = event.get("measuringPointCode", {})
+            if not event_type or not event_date or not measuring_point_code:
+                continue
+
+            event_types.append(event_type)
+            event_dates.append(event_date)
+            measuring_point_codes.append(measuring_point_code)
+
+        self.events_df = pl.DataFrame(
+            data={
+                "eventType": event_types,
+                "eventDate": event_dates,
+                "measuringPointCode": measuring_point_codes,
+            }
+        )
+        print(self.events_df)
 
     def _split_json_data(
         self, dispatch_document_data: dict[str, Any]
     ) -> tuple[dict[str, Any], list[dict[str, Any]]]:
         """Takes in the json data and splits it up into GMN and Measuringpoint data"""
+
+        intermediate_events = dispatch_document_data["GMN_PPO"].get(
+            "intermediateEvent", []
+        )
 
         measuringpoint_data = dispatch_document_data["GMN_PPO"].get(
             "measuringPoint", []
@@ -103,9 +135,10 @@ class GMNObjectImporter(ObjectImporter):
         gmn_data = dispatch_document_data["GMN_PPO"]
         gmn_data.pop("measuringPoint", None)
 
-        return gmn_data, measuringpoint_data
+        return gmn_data, measuringpoint_data, intermediate_events
 
     def _save_gmn_data(self, gmn_data: dict[str, Any]) -> None:
+        print(self.data_owner)
         self.gmn_obj = GMN.objects.update_or_create(
             bro_id=gmn_data.get("brocom:broId", None),
             data_owner=self.data_owner,
@@ -146,28 +179,44 @@ class GMNObjectImporter(ObjectImporter):
 
             # If a measuringpoint has old monitoringtube references, mp_data is a list
             # The last one is the active one, and therefore the one of interest
-            if isinstance(monitoring_tubes_data, list):
-                monitoring_tubes_data = monitoring_tubes_data[-1]
+            if isinstance(monitoring_tubes_data, dict):
+                monitoring_tubes_data = [monitoring_tubes_data]
 
-            monitoring_tube_data = monitoring_tubes_data.get(
-                "GroundwaterMonitoringTube", {}
-            )
-
-            Measuringpoint.objects.update_or_create(
-                gmn=self.gmn_obj,
-                data_owner=self.data_owner,
-                measuringpoint_code=mp_data.get("measuringPointCode", None),
-                defaults={
+            for monitoring_tube_reference in monitoring_tubes_data:
+                monitoring_tube_data = monitoring_tube_reference.get(
+                    "GroundwaterMonitoringTube", {}
+                )
+                event_date = monitoring_tube_data.get("startDate", {}).get(
+                    "brocom:date", None
+                )
+                mp_code = mp_data.get("measuringPointCode", None)
+                bro_id = monitoring_tube_data.get("broId", None)
+                print(f"{event_date} and {mp_code}")
+                event = self.events_df.filter(
+                    pl.col("eventDate").eq(event_date)
+                    & pl.col("measuringPointCode").eq(mp_code)
+                )
+                print(event)
+                if not event.is_empty():
+                    event_type = event.item(0, 0)
+                else:
+                    event_type = "GMN_StartRegistration"
+                defaults = {
                     "measuringpoint_start_date": mp_data.get("startDate", {}).get(
                         "brocom:date", None
                     ),
-                    "gmw_bro_id": monitoring_tube_data.get("broId", None),
                     "tube_number": monitoring_tube_data.get("tubeNumber", None),
-                    "tube_start_date": monitoring_tube_data.get("startDate", None).get(
-                        "brocom:date", None
-                    ),
-                },
-            )
+                    "tube_start_date": event_date,
+                    "event_type": event_type,
+                }
+
+                Measuringpoint.objects.update_or_create(
+                    gmn=self.gmn_obj,
+                    data_owner=self.data_owner,
+                    measuringpoint_code=mp_code,
+                    gmw_bro_id=bro_id,
+                    defaults=defaults,
+                )
 
 
 class GMWObjectImporter(ObjectImporter):