+++ gmn/choices.py
@@ -1,0 +1,6 @@
+GMN_EVENT_TYPES = (
+    ("GMN_MeasuringPoint", "GMN_MeasuringPoint"),
+    ("GMN_TubeReference", "GMN_TubeReference"),
+    ("GMN_MeasuringPointEndDate", "GMN_MeasuringPointEndDate"),
+    ("GMN_StartRegistration", "GMN_StartRegistration"),
+)