+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.18 (unreleased)
 
 -   Enhancement: Add request-types to bulkuploads.
+-   Enhancement: Improve GMW UploadModels
 
 ## 1.17 (2025-02-27)