+++ CHANGES.md
@@ -2,9 +2,8 @@
 
 ## 1.18 (unreleased)
 
-
-- Nothing changed yet.
-
+-   Enhancement: Add request-types to bulkuploads.
+-   Enhancement: Improve GMW UploadModels
 
 ## 1.17 (2025-02-27)